feat(schedule): 优化排班首页功能
- 添加获取员工项目信息的功能
- 重构添加排班和删除排班的逻辑
- 优化排班首页界面和交互
- 新增排班模态框管理钩子

+++ src/components/schedule/hooks/use-delete-schedule.tsx
@@ -1,19 +1,32 @@
-import {useDemeterDispatch} from "@D/core/store/demeter-hook";
+import {Modal} from "antd";
 import {useCallback} from "react";
+import {useDemeterDispatch} from "@D/core/store/demeter-hook";
 import {ProjectService} from "@D/core/service/project-service";
 import {ProjectEntity} from "@D/core/entity/project-entity";
 import {deleteProjectEntity} from "@D/core/store/features/project-slice";
 import {useAntdMessage} from "@D/core/hooks/use-antd-message";
 
 export const useDeleteSchedule = () => {
+    const [modal, contextHolder] = Modal.useModal();
     const dispatch = useDemeterDispatch();
     const {contextHolderMessage, success, failure} = useAntdMessage();
     const deleteSchedule = useCallback((projectId: string) => {
         const projectService = ProjectService.getInstance();
-        projectService.deleteProjectByIdRequest(projectId, (deletedProject: ProjectEntity) => {
-            dispatch(deleteProjectEntity(deletedProject));
-            success("Delete project successfully").then();
-        }, (error: Error) => failure(projectService.parseResponseError(error)));
-    }, [dispatch, failure, success]);
-    return {deleteScheduleHolderMessage: contextHolderMessage, deleteSchedule}
+        modal.confirm({
+            title: "Delete Schedule",
+            content: "Are you sure to delete this schedule?",
+            okText: "Yes",
+            cancelText: "No",
+            onOk: () => {
+                projectService.deleteProjectByIdRequest(projectId, (deletedProject: ProjectEntity) => {
+                    dispatch(deleteProjectEntity(deletedProject));
+                    success("Delete project successfully").then();
+                }, (error: Error) => failure(projectService.parseResponseError(error)));
+            }
+        })
+    }, [dispatch, failure, modal, success]);
+    return {
+        deleteScheduleHolderMessage: <>{contextHolder}{contextHolderMessage}</>,
+        deleteSchedule
+    }
 }
